Fix code coverage by coverals.io
Every aspect of the build should be uniform and universally executable.
Replacing the shell script with a maven plugin should achieve this.

Coverage is calculated only after successful build to ensure that the
tests pass without the instrumentation installed by cobertura.

The examples section is excluded. While examples execute code they
do not properly test it.

Clojure is excluded because it does not quite follow the class to
source naming conventions of java. Cobertura relies on these to
find the right source for each class